+++ sdk/face/azure-ai-vision-face/samples/sample_authentication.py
@@ -44,12 +44,8 @@ from shared.helpers import beautify_json, get_logger
 class FaceAuthentication:
     def __init__(self):
         load_dotenv(find_dotenv())
-        self.endpoint = os.getenv(
-            CONFIGURATION_NAME_FACE_API_ENDPOINT, DEFAULT_FACE_API_ENDPOINT
-        )
-        self.key = os.getenv(
-            CONFIGURATION_NAME_FACE_API_ACCOUNT_KEY, DEFAULT_FACE_API_ACCOUNT_KEY
-        )
+        self.endpoint = os.getenv(CONFIGURATION_NAME_FACE_API_ENDPOINT, DEFAULT_FACE_API_ENDPOINT)
+        self.key = os.getenv(CONFIGURATION_NAME_FACE_API_ACCOUNT_KEY, DEFAULT_FACE_API_ACCOUNT_KEY)
         self.logger = get_logger("sample_authentication")
 
     def authentication_by_api_key(self):
@@ -58,14 +54,12 @@ class FaceAuthentication:
         from azure.ai.vision.face.models import FaceDetectionModel, FaceRecognitionModel
 
         self.logger.info("Instantiate a FaceClient using an api key")
-        with FaceClient(
-            endpoint=self.endpoint, credential=AzureKeyCredential(self.key)
-        ) as face_client:
+        with FaceClient(endpoint=self.endpoint, credential=AzureKeyCredential(self.key)) as face_client:
             sample_file_path = helpers.get_image_path(TestImages.DEFAULT_IMAGE_FILE)
             result = face_client.detect(
                 helpers.read_file_content(sample_file_path),
-                detection_model=FaceDetectionModel.DETECTION_03,
-                recognition_model=FaceRecognitionModel.RECOGNITION_04,
+                detection_model=FaceDetectionModel.DETECTION03,
+                recognition_model=FaceRecognitionModel.RECOGNITION04,
                 return_face_id=False,
             )
 
@@ -80,14 +74,12 @@ class FaceAuthentication:
         from azure.ai.vision.face.models import FaceDetectionModel, FaceRecognitionModel
 
         self.logger.info("Instantiate a FaceClient using a TokenCredential")
-        with FaceClient(
-            endpoint=self.endpoint, credential=DefaultAzureCredential()
-        ) as face_client:
+        with FaceClient(endpoint=self.endpoint, credential=DefaultAzureCredential()) as face_client:
             sample_file_path = helpers.get_image_path(TestImages.DEFAULT_IMAGE_FILE)
             result = face_client.detect(
                 helpers.read_file_content(sample_file_path),
-                detection_model=FaceDetectionModel.DETECTION_03,
-                recognition_model=FaceRecognitionModel.RECOGNITION_04,
+                detection_model=FaceDetectionModel.DETECTION03,
+                recognition_model=FaceRecognitionModel.RECOGNITION04,
                 return_face_id=False,
             )