+++ sdk/face/azure-ai-vision-face/tests/test_authentication_async.py
@@ -28,8 +28,8 @@ class TestAuthenticationAsync(AzureRecordedTestCase):
         sample_file_path = helpers.get_image_path(TestImages.IMAGE_DETECTION_1)
         result = await client.detect(
             helpers.read_file_content(sample_file_path),
-            detection_model=FaceDetectionModel.DETECTION_03,
-            recognition_model=FaceRecognitionModel.RECOGNITION_04,
+            detection_model=FaceDetectionModel.DETECTION03,
+            recognition_model=FaceRecognitionModel.RECOGNITION04,
             return_face_id=False,
             return_face_attributes=[
                 FaceAttributeTypeDetection03.HEAD_POSE,